move common enums from proto to propriate folder
